# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/company_industries:
    get:
      tags:
        - CompanyIndustries
      summary: List CompanyIndustries
      description: Get a paginated list of CompanyIndustries with optional filtering and sorting
      operationId: listCompanyIndustries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyIndustryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CompanyIndustries
      summary: Create CompanyIndustry
      description: Create a new CompanyIndustry
      operationId: createCompanyIndustry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyIndustryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyIndustry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/departments:
    get:
      tags:
        - Departments
      summary: List Departments
      description: Get a paginated list of Departments with optional filtering and sorting
      operationId: listDepartments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Departments
      summary: Create Department
      description: Create a new Department
      operationId: createDepartment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    CompanyIndustry:
      type: object
      required:
        - id
        - company_id
        - industry_id
        - is_primary
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        industry_id:
          type: string
          format: uuid
        is_primary:
          type: boolean
        metadata:
          type: object
...
    CreateCompanyIndustryRequest:
      type: object
      required:
        - company_id
        - industry_id
        - is_primary
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        industry_id:
          type: string
          format: uuid
        is_primary:
          type: boolean
        metadata:
          type: object
...
    CompanyIndustryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CompanyIndustry'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    Department:
      type: object
      required:
        - id
        - company_id
        - code
        - name
        - level
        - is_group
        - sort_order
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 150
        parent_id:
          nullable: true
          type: string
          format: uuid
        level:
          type: integer
          format: int32
        is_group:
          type: boolean
        branch_id:
          nullable: true
          type: string
          format: uuid
        manager_id:
          nullable: true
          type: string
          format: uuid
        sort_order:
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    CreateDepartmentRequest:
      type: object
      required:
        - company_id
        - code
        - name
        - level
        - is_group
        - sort_order
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 150
        parent_id:
          nullable: true
          type: string
          format: uuid
        level:
          type: integer
          format: int32
        is_group:
          type: boolean
        branch_id:
          nullable: true
          type: string
          format: uuid
        manager_id:
          nullable: true
          type: string
          format: uuid
        sort_order:
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    DepartmentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Department'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []